Updarte before rewrite of DAG objects and addition of context manager.  Also, push Spacial Transformer and Cyclic Transformer classes.

+++ utility.cpp
@@ -79,8 +79,8 @@ void resize_image(const Matrix& input_image, const float scale_factor, Matrix& o
       for (int x = 0; x < new_width; ++x)
       {
          // Calculate the corresponding pixel position in the input image
-         const float input_x = static_cast<float>(x) / scale_factor;
-         const float input_y = static_cast<float>(y) / scale_factor;
+         const double input_x = static_cast<float>(x) / scale_factor;
+         const double input_y = static_cast<float>(y) / scale_factor;
 
          // Compute the pixel values of the four nearest neighbors
          const int x0 = static_cast<int>(input_x);
@@ -88,22 +88,23 @@ void resize_image(const Matrix& input_image, const float scale_factor, Matrix& o
          const int x1 = std::min(x0 + 1, static_cast<int>(input_image.cols() - 1));
          const int y1 = std::min(y0 + 1, static_cast<int>(input_image.rows() - 1));
 
-         const float p00 = input_image(y0, x0);
-         const float p01 = input_image(y0, x1);
-         const float p10 = input_image(y1, x0);
-         const float p11 = input_image(y1, x1);
+         const double p00 = input_image(y0, x0);
+         const double p01 = input_image(y0, x1);
+         const double p10 = input_image(y1, x0);
+         const double p11 = input_image(y1, x1);
 
          // Compute the bilinear interpolation weights
-         const float weight_x1 = input_x - static_cast<float>(x0);
-         const float weight_x0 = 1.0f - weight_x1;
-         const float weight_y1 = input_y - static_cast<float>(y0);
-         const float weight_y0 = 1.0f - weight_y1;
+         const double weight_x1 = input_x - static_cast<float>(x0);
+         const double weight_x0 = 1.0f - weight_x1;
+         const double weight_y1 = input_y - static_cast<float>(y0);
+         const double weight_y0 = 1.0f - weight_y1;
 
          // Compute the interpolated pixel value
-         const float interpolated_value = weight_x0 * (weight_y0 * p00 + weight_y1 * p10) + weight_x1 * (weight_y0 * p01 + weight_y1 * p11);
+         const double interpolated_value = weight_x0 * (weight_y0 * p00 + weight_y1 * p10) + weight_x1 * (weight_y0 * p01 + weight_y1 * p11);
 
          // Set the output pixel value
          output_image(y, x) = interpolated_value;
       }
    }
 }
+
